refactor: DB 구조 변경에 따른 Users, UsersService, UsersController 수정
